Move LocalStackTestBase to localstack-test module
