+++ java/common/localstack-test/src/main/java/sleeper/localstack/test/LocalStackTestBase.java
@@ -13,21 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.configuration.testutils;
+package sleeper.localstack.test;
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import com.amazonaws.services.sqs.AmazonSQS;
+import com.amazonaws.services.sqs.AmazonSQSClientBuilder;
 import org.testcontainers.containers.localstack.LocalStackContainer;
 import org.testcontainers.junit.jupiter.Testcontainers;
-
-import sleeper.localstack.test.SleeperLocalStackContainer;
 
 import static sleeper.localstack.test.LocalStackAwsV1ClientHelper.buildAwsV1Client;
 
 /**
- * A base class for tests to run against LocalStack in this module.
+ * A base class for tests to run against LocalStack.
  */
 @Testcontainers
 public abstract class LocalStackTestBase {
@@ -36,5 +36,6 @@ public abstract class LocalStackTestBase {
 
     protected final AmazonS3 s3Client = buildAwsV1Client(CONTAINER, LocalStackContainer.Service.S3, AmazonS3ClientBuilder.standard());
     protected final AmazonDynamoDB dynamoDBClient = buildAwsV1Client(CONTAINER, LocalStackContainer.Service.DYNAMODB, AmazonDynamoDBClientBuilder.standard());
+    protected final AmazonSQS sqsClient = buildAwsV1Client(CONTAINER, LocalStackContainer.Service.SQS, AmazonSQSClientBuilder.standard());
 
 }
